Add support for setting environment variables as "common" options
Setting environment variables like this is useful for cross-compilation builds
(much like setting the `target_platform`!), and should allow mopack to do a
better job of finding packages. In the short term, however, this only applies
to Boost packages.

This also renames "general options" to "common options" and "global options"
to "this options" (which isn't very good grammar, but that's ok).