Initial support for conditional packages
Package definitions can now be a list, with each item containing an `if`
directive (the last item can exclude the `if`). The first definition
satisfying the conditional will be used. Currently, however, the `if`
directive is just a simple boolean value. Later, we'll turn this into
something that can be evaluated to do interesting things, like platform-
specific package definitions.