Improve how default package settings are inherited
This changes the behavior to only inherit defaults if the config doesn't
specify any of its own parameters, or if `inherit_defaults` is true. This
makes it easier to override things without having to explicitly set every
parameter.

It might even be useful to disable inheriting defaults when the config has no
parameters, and provide some more ways to opt into the behavior.